A debugRenderer legalább működik

+++ core/src/hu/wolfman/deimos/Constants.java
@@ -8,7 +8,7 @@ public class Constants {
     public static final int WIDTH = 560;
     public static final int HEIGHT = 280;
     
-    public static final int FPS = 60;
+    public static final float FPS = 60;
     public static final float STEP = 1/FPS;
     
     public static final String MAINGAME = "main";
